requirements.txt
pycryptodome==3.7.3
requests==2.32.3
pytz